# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the BuffetPOS API documentation.",
        "title": "BuffetPOS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/loyalty/add-point": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add points to a customer's account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loyalty"
                ],
                "summary": "Add Points to Customer",
                "parameters": [
                    {
                        "description": "Add Point Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CustomerAddPointRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BaseCustomer"
                        }
                    }
                }
            }
        },
        "/loyalty/customer/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a customer by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loyalty"
                ],
                "summary": "Delete a Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/loyalty/customers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find all customers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loyalty"
                ],
                "summary": "Find All Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.BaseCustomer"
                            }
                        }
                    }
                }
            }
        },
        "/loyalty/redeem": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Redeem points from a customer's account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loyalty"
                ],
                "summary": "Redeem Points from Customer",
                "parameters": [
                    {
                        "description": "Redeem Point Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CustomerRedeemRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BaseCustomer"
                        }
                    }
                }
            }
        },
        "/loyalty/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register a new customer in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loyalty"
                ],
                "summary": "Register a New Customer",
                "parameters": [
                    {
                        "description": "Register Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CustomerRegisterRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CustomerAddPointRequest": {
            "type": "object",
            "required": [
                "phone",
                "pin",
                "point"
            ],
            "properties": {
                "phone": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                }
            }
        },
        "requests.CustomerRedeemRequest": {
            "type": "object",
            "required": [
                "invoice_id",
                "phone",
                "pin"
            ],
            "properties": {
                "invoice_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                }
            }
        },
        "requests.CustomerRegisterRequest": {
            "type": "object",
            "required": [
                "phone",
                "pin"
            ],
            "properties": {
                "phone": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                }
            }
        },
...
        "responses.BaseCustomer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                }
            }
        },
...
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}